// src/app/models/reclamation.ts
export interface Reclamation {
  id?: number;
  dossier: string;
  date_reclamation: string;
  Nom_station: string;
  Secteur: string;
  N_Arbre: string;
  Demandeur: string;
  Numero_courrier: string;
  date_reception: string;
  reclamation: string;
  decision: string;
  date_decision: string;
  etat: boolean;
}